modularized Model
